Make implementations of add/delete_constraint take unsigned int
Before this, the declaration was changed to unsigned int for
add_constraint and delete_constraint, but the implementation used
an `int&`. This is especially bad for `delete_constraint` as it
used a variable that needs to be unsigned to -1 to indicate "not
found". It did then however not use this to check if the constraint to
delete was found - instead it would exhibit some unknown behavior caused
by a -1 being a very large number.